+++ src/Modules/PlayerRecords/Events/PbRecordUpdateEventArgs.cs
@@ -1,7 +1,7 @@
 ﻿using EvoSC.Common.Interfaces.Models;
 using EvoSC.Modules.Official.PlayerRecords.Interfaces.Models;
 
-namespace EvoSC.Modules.Official.PlayerRecords;
+namespace EvoSC.Modules.Official.PlayerRecords.Events;
 
 public class PbRecordUpdateEventArgs : EventArgs
 {
